requirements.txt
tensorflow==2.10.0
streamlit
numpy
matplotlib
pillow
